CI/CD: improve release flow and add separate PR comment flow (#418)
This pull request includes changes to the GitHub workflows in the
repository. The most important changes include adding a new workflow
that comments on a pull request with artifact details, adding conditions
to a workflow to execute steps only when a specific dependency is
present, and modifying steps in a workflow.

Main workflow changes:

* `.github/workflows/pr-comment.yml`: Added a new workflow that comments
on a pull request with artifact details when a specific workflow is
completed. (`.github/workflows/pr-comment.yml`)
* `.github/workflows/bump-pbiviz-on-dependabot.yml`: Added conditions to
the steps in the "bump-pbiviz-on-dependabot" workflow to only execute
them when the "powerbi-visuals-api" dependency is present.
(`.github/workflows/bump-pbiviz-on-dependabot.yml`)
* `.github/workflows/release.yml`: The "Commit changes" step in the
"release" workflow was removed and then added back with modifications.
(`.github/workflows/release.yml`)
* `.github/workflows/pr-build.yml`: Several steps were removed from the
"pr-build" workflow, possibly due to issues or redundancy.
(`.github/workflows/pr-build.yml`)